Simplify build process

+++ scripts/compile.sh
@@ -1,52 +1,67 @@
-#!/usr/bin/env bash
-#
-# This script builds the application from source for multiple platforms.
+#!/usr/bin/env sh
 set -e
 
-# Get the parent directory of where this script is.
-SOURCE="${BASH_SOURCE[0]}"
-while [ -h "$SOURCE" ] ; do SOURCE="$(readlink "$SOURCE")"; done
-DIR="$(cd -P "$(dirname "$SOURCE")/.." && pwd)"
+if [ -z "$NAME" ]; then
+  echo "Missing \$NAME!"
+  exit 127
+fi
 
-# Change into that directory
-cd "$DIR"
+if [ -z "$PROJECT" ]; then
+  echo "Missing \$PROJECT!"
+  exit 127
+fi
 
-# Get the name from the directory and package
-PKGNAME="$(basename $(cd ../../ && pwd))/$(basename $(cd ../ && pwd))/$(basename $(pwd))"
-NAME=${NAME:-"$(basename $(pwd))"}
+# Get the git commit information
+GIT_COMMIT="$(git rev-parse HEAD)"
+GIT_DIRTY="$(test -n "$(git status --porcelain)" && echo "+CHANGES" || true)"
 
-# Get the git commit
-GIT_COMMIT=$(git rev-parse HEAD)
-GIT_DIRTY=$(test -n "`git status --porcelain`" && echo "+CHANGES" || true)
-
-# Determine the arch/os combos we're building for
-XC_ARCH=${XC_ARCH:-"amd64"}
-XC_OS=${XC_OS:-"darwin linux windows"}
-XC_EXCLUDE=${XC_EXCLUDE:-""}
-
-# Delete the old dir
-echo "==> Removing old builds..."
-rm -f bin/*
+# Remove old builds
+rm -rf bin/*
 rm -rf pkg/*
-mkdir -p bin/
 
-# Build
-echo "==> Building..."
-export CGO_ENABLED=0
+# Runtime variables
+LDFLAGS="-s -w"
+LDFLAGS="$LDFLAGS -X main.Name=${NAME}"
+LDFLAGS="$LDFLAGS -X main.Version=${VERSION}"
+LDFLAGS="$LDFLAGS -X main.GitCommit=${GIT_COMMIT}${GIT_DIRTY}"
+
+# Build!
 for GOOS in $XC_OS; do
   for GOARCH in $XC_ARCH; do
-    if [[ $XC_EXCLUDE == *"${GOOS}/${GOARCH}"* ]]; then
+    if [ "$XC_EXCLUDE" = "*${GOOS}/${GOARCH}*" ]; then
       continue
     fi
 
-    printf "%s%20s %s\n" "-->" "${GOOS}/${GOARCH}:" "${PKGNAME}"
-    env \
+    printf "%s%20s %s\n" "-->" "${GOOS}/${GOARCH}:" "${PROJECT}"
+    env -i \
+      PATH="$PATH" \
+      CGO_ENABLED=0 \
+      GOPATH="$GOPATH" \
+      GOROOT="$GOROOT" \
       GOOS="${GOOS}" \
       GOARCH="${GOARCH}" \
       go build \
       -a \
-      -ldflags="-X main.GitCommit=${GIT_COMMIT}${GIT_DIRTY}" \
+      -ldflags="$LDFLAGS" \
       -o="pkg/${GOOS}_${GOARCH}/${NAME}" \
       .
   done
 done
+
+echo "--> Compressing..."
+mkdir pkg/dist
+for PLATFORM in $(find ./pkg -mindepth 1 -maxdepth 1 -type d); do
+  OSARCH=$(basename ${PLATFORM})
+  if [ "$OSARCH" = "dist" ]; then
+    continue
+  fi
+
+  cd $PLATFORM
+  tar -czf ../dist/${NAME}_${VERSION}_${OSARCH}.tgz ${NAME}
+  cd - >/dev/null 2>&1
+done
+
+echo "--> Checksumming..."
+cd pkg/dist
+shasum -a256 * > "${NAME}_${VERSION}_SHA256SUMS"
+cd - >/dev/null 2>&1
